vdom: fix column width
